Theme Upload: Move error handling to notices middleware
This also allows us to handle errors occurring when trying to install a WPCOM theme from the Theme Showcase.
